requirements.txt
numpy
pandas
openpyxl
Pillow
scikit-learn
torch
torchvision
matplotlib
